requirements.txt
python-telegram-bot==20.3
easyocr
pillow
requests
certifi
